First Version where i am able to see what all scintillators fire on the top plane. demo can be seen in TestVisualization.
This verison makes use of MultiThreading, where the main thread is used to display the
geometry and a child thread is spawned to modify the parameters for the scintillator which fires.

Still not taking data from Real file, instead generated random strip num and displaying
that scintillator in Red color, which show that particular scinitillator fires.

Later on this logic needs to be replaced with following logic
"If fScintHit is set then it signified that it fire and it should be displayed in RED color