Simulation modified so as to also store scattering angle in the ROOT file. Also checked 3DHitPlot of detectors. Observed significant change in pattern before and after introduction of scatterer.
To make sure that most of the muon passes through scatterer, the range of random
number generation along X and Y Direction is set to  -15cm to 15 cm, because the
dimension of scatterer is 30cm x 30cm x 30cm.

   In the experimental setup this may be achieved by putting the paddle just on top
of scatterer and then taking 3fold coincidence. Even better approach may be to put
two paddle, orthogonal to each other so it will further confine the muon and then
taking 4fold coincidence.