The basic version of Simulation is modified to use EventHelper and RunHelper, so that one can calculate the SD of scattering angle in each voxel.
CommonFunc class is now made as Singleton class. This is required so that we can call non static
functions also in it.

Curently we are using ROOTs histogram SD, Previously written SD function
was giving wrong hence it needs a relook.

Simulation2 is not yet tested again, which needs to be tested and fixed to give results comparable
to SimulationBasic.