Trying to implement new idea discussed with Tushar, to allocated the scattering density among all the enclosing voxel around the imaginary voxel centered at PocaPoint.
For the time being , the scattering density is divided among the
enclosing voxels in the inverse proportion of the distance between
enclosing voxel center and  Poca Point.

But finally the density should be distributed in the proportion of
the volume, that enclosing voxel shares with the imagingary voxel
centered around poca point.