New ideas implementation.
In addition to MLEM,
This we also have the stuff where we are trying to use Unscatttered track. NOT CONCLUSION YET.
Ideas implementation are in Progress.

NOW we need to dissect each and every concept that is implemented. Everything needs thorogh test.
May be the right time to do code REFACTORING.

for eg.
1) if there is not scatterer are we really able to get straight tracks.
2) Getting clean Voxel count from each Voxel
3) Clean Voxel check using vertical tracks without scatterer.
4) Testing MLEM, and detect where we are wrong. It gives something which is not consistent,
   atleast not till NOW.