Following important changes done that starts giving some positive results for "Sample Track from Pixel" idea.
A critical bug fix in Pixel class, where previously we were missed to
set the Z coordinate of returned random point. Now this is set equal to
the Z of Pixel.

Voxelator modified and a new function GetFilteredVoxelVectorBasedOnWeightedCount
is added which will try to find the filtered Voxels based on WeightedCount stored
in each Voxel, Currently if weighted count is greater than zero then we consider
it as signal voxel else background voxel and PoCA belong to background declared
voxels are ignored.