Added a FLAG in CMake to enable or disable single point muon calibration. By default set to ON
When ON it makes use of ** vecOfPeakPos ** defined in ** HardwareNomenclature **
to do single point calibration.
Hence if peaks are shifting then ** vecOfPeakPos ** needs to be updated to
have proper energy calibration for Muon.

** TestScintillatorPeaks ** will show energy spectrum of all the scintillators,
and can be used to get the peak position to be put in ** vecOfPeakPos **.

Ideally the logic to calculate the PeakPos should be automatic, but currently
it is manually checked and filled in ** vecOfPeakPos **.