Good progress, atleast now able to see the scintillators fingers for all the top three layers. for layer8 first version of function in ml.cpp is working nicely.
Third version is written to work for all the combination, but unfortunately does not work
for Cross layer, but work nicely for oblong layers of TOP set of scintillators.
This needs to be fixed.

But in lower layers, not able to see scintillator finger pattern from experimental data,
and obviously seen from simulated data.