A lot a work done to get the improved resolution. Tried by building regression model for whole bar, but then it gives skewed prediction histograms, which signifies some correction needs to applied at lower level.
Then also tried a Hybrid approach, where a classification model is build
to detect the class (source postion) to which a sample belongs to,
After that regression model for that particlular class is used to
predict the Z. Giving good results for characterization if using
datasets for all the position.

But if we exclude some datasets from classification model, then
then obviously we are getting wrong classification and hence wrong
Z.

PS :None of the schemes are tested with Muon. Currently all the testing
    work is done with gammas